+++ packages/eyes-sdk-core/lib/fluent/DriverCheckSettings.js
@@ -995,8 +995,10 @@ class CheckSettings {
     async function persistRegions(regions = []) {
       const persisted = []
       for (const region of regions) {
-        persisted.push(await region.toPersistedRegions(driver))
+        const persistedRegions = await region.toPersistedRegions(driver)
+        persisted.push(...persistedRegions)
       }
+      return persisted
     }
   }
 }