+++ packages/eyes-webdriverio-5-service/CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Unreleased
 
 
+## 1.6.6 - 2020/7/5
+
+- updated to @applitools/eyes-webdriverio@5.17.0 (from 5.16.0)
+
 ## 1.6.5 - 2020/7/1
 
 - updated to @applitools/eyes-webdriverio@5.16.0 (from 5.15.1)