+++ packages/eyes-webdriverio-5/CHANGELOG.md
@@ -6,6 +6,11 @@
 - updated to @applitools/eyes-sdk-core@11.2.1 (from 11.2.2)
 - updated to @applitools/visual-grid-client@14.5.1 (from 14.5.2)
 
+## 5.17.0 - 2020/7/5
+
+- updated to @applitools/eyes-sdk-core@11.2.2 (from 11.2.1)
+- updated to @applitools/visual-grid-client@14.5.2 (from 14.5.1)
+
 ## 5.16.0 - 2020/6/30
 
 - updated to @applitools/eyes-sdk-core@11.2.0 (from 11.1.0)