+++ packages/visual-grid-client/CHANGELOG.md
@@ -3,6 +3,16 @@
 ## Unreleased
 
 
+## 14.5.3 - 2020/7/7
+
+- updated to @applitools/eyes-sdk-core@11.3.0 (from 11.2.2)
+
+## 14.5.2 - 2020/7/5
+
+- fetch resources with tunneling agent if necessary  ([Trello](https://trello.com/c/a22B3tUB))
+- updated to @applitools/dom-snapshot@3.6.1 (from 3.6.0)
+- updated to @applitools/eyes-sdk-core@11.2.2 (from 11.2.1)
+
 ## 14.5.1 - 2020/6/30
 
 - updated to @applitools/eyes-sdk-core@11.2.0 (from 11.1.0)